Facinating Number Check Using Method 2
To Check Wether the given num is Facinating Num or Not Using String
Hint: A Fascinating Number is a number that, when multiplied by 2 and 3, and the
      results are concatenated with the original number, contains all digits from 1 to 9 exactly once (excluding 0).
Ex:
Take an input number (num). 192
Multiply num by 2 and 3, then concatenate these values with num.
Check if the concatenated number contains all digits (1-9) exactly once. i.e frequncy of all digits
If all digits are present exactly once, print "Fascinating Number!", else print "Not a Fascinating Number!".