PhloemNumCheck
A Phloem Number is a number where the sum of the first and last digit is not equal to the sum of the middle digits.

Example:
Consider 101

First + Last Digit = 1 + 1 = 2
Middle Digit Sum = 0
Since 2 ≠ 0, 101 is a Phloem Number
Phloem numbers are simply the opposite of Xylem Numbers, meaning they do not satisfy the Xylem condition.