Armstrong Num Range
Check the Armstrong num in user entered range using recusion Method

An Armstrong Number (also called a Narcissistic Number) is a number where the sum of its digits each raised to the power of the total number of digits equals the number itself.
Example:
153 → 1^3+5^3+3^3=1+125+27=153 ✅
If a number satisfies this condition, it is an Armstrong Number. 🚀Arm