Static Variable Access from the Another Class
To Access the Static Var from another class to current class.

Note: When we access a static var form another class then it prints the it's initial value that var has to assigned, the static var has always assigned by default with it's initial value or defualt value. so it print that value but if we already assign some value to the static var so it is it's initial value so it gives that value.